refactor(store): move message-related modules and types to dedicated messages directory
Restructure the store module by extracting all message-related functionality into a dedicated messages directory. This includes:
- Moving message types and related structs to messages/types.rs
- Creating separate modules for message history, threads, agents, cursors, and targets
- Relocating AgentConfig from store/agents.rs to agent/config.rs
- Updating imports across the codebase to reflect new module structure

The changes improve code organization and maintainability while keeping all existing functionality intact.
