+++ src/agent/drivers/claude.rs
@@ -2,9 +2,10 @@ use std::io::Write as _;
 use std::process::{Child, Command, Stdio};
 
 use super::{command_exists, run_command, Driver, ParsedEvent, SpawnContext};
+use crate::agent::config::AgentConfig;
 use crate::agent::drivers::prompt::{build_base_system_prompt, PromptOptions};
 use crate::agent::runtime_status::{RuntimeAuthStatus, RuntimeStatus};
-use crate::store::agents::{AgentConfig, AgentRuntime};
+use crate::store::agents::AgentRuntime;
 
 pub struct ClaudeDriver;
