+++ src/agent/drivers/kimi.rs
@@ -3,9 +3,10 @@ use std::path::Path;
 use std::process::{Child, Command, Stdio};
 
 use super::{command_exists, home_dir, read_file, Driver, ParsedEvent, SpawnContext};
+use crate::agent::config::AgentConfig;
 use crate::agent::drivers::prompt::{build_base_system_prompt, PromptOptions};
 use crate::agent::runtime_status::{RuntimeAuthStatus, RuntimeStatus};
-use crate::store::agents::{AgentConfig, AgentRuntime};
+use crate::store::agents::AgentRuntime;
 
 pub struct KimiDriver;
