+++ src/agent/mod.rs
@@ -1,5 +1,6 @@
 pub mod activity_log;
 pub mod collaboration;
+pub mod config;
 pub mod drivers;
 pub mod lifecycle;
 pub mod manager;
