+++ src/store/events.rs
@@ -4,8 +4,8 @@ use rusqlite::{params, Connection, Transaction};
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
-use crate::utils::parse_datetime;
 use super::Store;
+use crate::utils::parse_datetime;
 
 /// One durable IM event persisted for replay and realtime fanout.
 #[derive(Debug, Clone, Serialize, Deserialize)]
