+++ tests/check_impl.rs
@@ -1,0 +1,9 @@
+struct MyStruct;
+mod a {
+    impl super::MyStruct {
+        pub fn foo() {}
+    }
+}
+fn main() {
+    MyStruct::foo();
+}
