+++ tests/driver_tests.rs
@@ -1,8 +1,8 @@
+use chorus::agent::config::AgentConfig;
 use chorus::agent::drivers::claude::ClaudeDriver;
 use chorus::agent::drivers::codex::CodexDriver;
 use chorus::agent::drivers::kimi::KimiDriver;
 use chorus::agent::drivers::Driver;
-use chorus::store::agents::AgentConfig;
 
 #[test]
 fn test_claude_prompt_uses_split_message_tools() {
